+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -536,6 +536,13 @@ def validate_assign_kubelet_identity(namespace):
                 "--assign-kubelet-identity is not a valid Azure resource ID.")
 
 
+def validate_prompt_input(namespace):
+    if namespace.prompt is None:
+        return
+    if not re.search(r'[a-zA-Z]', namespace.prompt):
+        raise InvalidArgumentValueError('--prompt does not contain any alphabet character')
+
+
 def validate_snapshot_name(namespace):
     """Validates a nodepool snapshot name to be alphanumeric and dashes."""
     rfc1123_regex = re.compile(