+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,14 @@ Release History
 ===============
 upcoming
 ++++++
+* Add command group 'az containerapp connected-env', support show/list/delete/create connected environment
+
+0.3.39
+++++++
+* 'az containerapp update': fix bug for populating secret value with --yaml
+
+0.3.38
+++++++
 * Add support for binding managed MySQL Flexible server to a containerapp
 * Removed preview tag for some command groups and params (e.g. 'az containerapp job', 'az containerapp env storage', 'az containerapp env workload-profile')
 * 'az containerapp env': --enable-workload-profiles allowed values:true, false